+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformBasePlugin.kt
@@ -446,7 +446,7 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
                 is SigningAware,
                 is SplitModeAware,
                 is TestableAware,
-                -> action()
+                    -> action()
             }
         }
 }