+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformModulePlugin.kt
@@ -25,8 +25,10 @@ abstract class IntelliJPlatformModulePlugin : Plugin<Project> {
 
         listOf(
             // Build Module
+            GenerateManifestTask,
             JarCompanion,
             InstrumentCodeTask,
+            InstrumentedJarTask,
             PrepareSandboxTask,
 
             // Test Module